+++ src/Schema/FloatType-7.0.php
@@ -53,15 +53,15 @@ class FloatType extends AbstractType
             'type' => T::Enum(['float']),
             'args' => T::Struct([
                 'nullable'  => T::Bool(T::NULLABLE),
-                'min'       => T::Int(T::NULLABLE),
-                'max'       => T::Int(T::NULLABLE),
+                'min'       => T::Float(T::NULLABLE),
+                'max'       => T::Float(T::NULLABLE),
             ])
         ]);
     }
 
     public function getErrors($value)
     {
-        if ($error = parent::getErrors($value)){
+        if ($error = parent::getErrors($value)) {
             return $error;
         }
